+++ src/main/java/com/botapeer/domain/service/interfaces/IPlantRecordService.java
@@ -6,7 +6,7 @@ import java.util.Optional;
 import com.botapeer.domain.model.plantRecord.PlantRecord;
 
 public interface IPlantRecordService {
-	public Optional<PlantRecord> findById(int plantId);
+	public Optional<PlantRecord> findById(Long plantId);
 
 	public Optional<PlantRecord> create(PlantRecord plantRecord);
