+++ src/main/java/com/botapeer/infrastructure/repository/PostRepositoryImpl.java
@@ -19,7 +19,7 @@ public class PostRepositoryImpl implements IPostRepository {
 	private final PostMapper postMapper;
 
 	@Override
-	public Collection<Post> findByPlantRecordId(int id) {
+	public Collection<Post> findByPlantRecordId(Long id) {
 		Collection<PostEntity> entity = postMapper.findByPlantRecordId(id);
 		Collection<Post> record = PostRepositoryDto.toModel(entity);
 		return record;
